requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic[email]>=2.9.0
pydantic-settings>=2.5.0

# Database
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
alembic>=1.14.0

# Redis
redis[hiredis]>=5.2.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0,<4.0.0
cryptography>=43.0.0

# LLM / Agents
openai>=1.55.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# HTTP / OAuth
httpx>=0.27.0
authlib>=1.3.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.4.0
python-multipart>=0.0.12
